+++ CHANGELOG.md
@@ -3,6 +3,7 @@ x.y.z / 2014-06-dd
 - removed semicolons per http://cryptocoinjs.com/about/contributing/#semicolons
 - removed `ECKey` dep
 - added `ecurve` dep
+- removed `terst` dev dep for `assert`
 
 0.0.1 / 2014-05-29
 ------------------